+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/models/Config.kt
@@ -2,9 +2,15 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.base.models
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.ListSerializer
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
-import com.github.insanusmokrassar.TelegramBotAPI.types.ChatId
-import com.github.insanusmokrassar.TelegramBotAPI.types.toChatId
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.*
+import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.UpdatesFilter
+import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.startGettingOfUpdates
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.channels.SendChannel
 import kotlinx.serialization.*
 import org.h2.Driver
 
@@ -52,15 +58,54 @@ class Config (
             (logsChatId ?: sourceChatId).toChatId(),
             botConfig.createBot(),
             databaseConfig,
-            plugins
+            plugins,
+            botConfig.webhookConfig
         )
 }
 
-class FinalConfig (
+data class FinalConfig (
     val targetChatId: ChatId,
     val sourceChatId: ChatId,
     val logsChatId: ChatId,
     val bot: RequestsExecutor,
     val databaseConfig: DatabaseConfig,
-    val pluginsConfigs: List<Plugin> = emptyList()
-)
+    val pluginsConfigs: List<Plugin> = emptyList(),
+    private val webhookConfig: WebhookConfig?
+) {
+    fun createFilter(
+        messagesChannel: SendChannel<MessageUpdate>,
+        channelPostChannel: SendChannel<ChannelPostUpdate>,
+        mediaGroupChannel: SendChannel<List<MediaGroupUpdate>>,
+        callbackQueryChannel: SendChannel<CallbackQueryUpdate>
+    ): UpdatesFilter = UpdatesFilter(
+        {
+            messagesChannel.send(it)
+        },
+        {
+            mediaGroupChannel.send(it)
+        },
+        channelPostCallback = {
+            channelPostChannel.send(it)
+        },
+        channelPostMediaGroupCallback = {
+            mediaGroupChannel.send(it)
+        },
+        callbackQueryCallback = {
+            callbackQueryChannel.send(it)
+        },
+        editedMessageMediaGroupCallback = null,
+        editedChannelPostMediaGroupCallback = null
+    )
+
+    suspend fun subscribe(filter: UpdatesFilter, scope: CoroutineScope = NewDefaultCoroutineScope(4)) {
+        webhookConfig ?.setWebhook(
+            bot,
+            filter,
+            scope
+        ) ?: bot.startGettingOfUpdates(
+            scope = scope,
+            allowedUpdates = filter.allowedUpdates,
+            block = filter.asUpdateReceiver
+        )
+    }
+}