+++ CHANGELOG.md
@@ -302,3 +302,4 @@ available messages for post
 
 * Update version of `TelegramBotAPI` from `0.11.0` to `0.12.0`
 * Remove `CONFLATED_MODE_ON` environment variable checking. Now all channels by default conflated
+* Added section to bot config `webhookConfig`